refactor(storage): give serialization.py's codec a real ADR-061 owner
Closes ADR-061 gap E/B's last open item: the ~1500-line snapshot_to_dict/
snapshot_from_dict codec proper (and its schema-version history) moves
out of abicheck/serialization.py into a real storage-classified home.

- New abicheck/storage/snapshot_codec.py owns decode_snapshot/
  finalize_snapshot/load_snapshot_document/save_snapshot/write_snapshot,
  plus platform-block/provenance decode and the AbiSnapshot(...) assembly.
- Split into siblings purely to stay under the ADR-061 new-file
  production line ceiling (a brand-new file gets no adoption-debt
  exemption): snapshot_schema_versions.py (SCHEMA_VERSION history/
  thresholds), snapshot_encode.py (encode direction),
  snapshot_decode_declarations.py (functions/variables/types/enums/
  typedefs decode), snapshot_reliability_flags.py (the seven
  *_facts_reliable flag computations).
- abicheck/serialization.py is now a ~280-line, delegation-only facade.
  It stays permanently unclassified (public_root_surfaces), confirmed
  the same "no single layer" way as checker_policy/contract_gating/
  reclassify: it is the one legal route through which
  workflows.snapshot_load.backfill_python_ext_from_evidence and
  policy.analysis_assurance_degraded_facts.degraded_reliability_facts
  run between decode_snapshot and finalize_snapshot -- neither call is
  legal from a storage-classified module (may_import: [model] only).
- decode_snapshot takes an explicit max_known_schema_version parameter
  so serialization.SCHEMA_VERSION (patchable in tests) is still the
  value actually consulted, now that the constant's real home moved out
  of the facade module.
- The one real serialization.py <-> storage.bundle_facts_codec cycle is
  unchanged, still resolved dynamically via importlib.import_module, not
  a new IMPORT_CYCLE_ALLOWLIST entry.
- Updated ADR-061's own text, known-gaps.md, and architecture/debt.yaml
  (removed the now-closed abicheck/serialization.py no-growth entry) to
  record the closure; tests/unit/storage/adr062_scope.py's
  NON_ADR062_MODULES gained the five new module names.

Every documented Python API name (load_snapshot/save_snapshot/
write_snapshot/snapshot_to_dict/snapshot_from_dict/snapshot_to_json/
snapshot_content_digest/SCHEMA_VERSION/the bundle_facts_* helpers) keeps
its historical signature and import path -- no behavior change.

Verified: mypy (759 files, 0 errors), ruff check/format, full
check_architecture.py and check_ai_readiness.py runs (identical to the
pre-change baseline, no new findings), and ~3700+ targeted fast tests
(serialization/schema-compat/bundle-facts/storage/buildsource/release/
python-ext and friends) all passing.
